Correct bug where clicking on a clist row does not always send you to the correct record
When a record is modified, and a user clicks on a different row in a clist instead of hitting
the 'apply changes' button`, and the user selects that the changes be saved then there is the
possibility that the entry that the user selected will no longer be in the row that the user selected.
This is caused by the addition of the new or modified record and the subsequent default sorting of the
clist.  The bug is corrected by finding out which record, rather than which row, the user was trying to
access and then going to that record after the new or modified one has been added.