Work around GTK bug in nested callback routines.
When creating a new appt in the datebook the first time the pulldown menus
for setting the time are used they appear and disappear.  One has to click
a second time for the menus to stay on the screen.  In addition, if one
clicks and holds the mouse the very first time then the entry for 8 am is
missing in the menu.  This patch fixes both items.