Use a slightly more robust scheme to decide when Mail versus Dial button should be displayed
Old scheme relied on the gettext translation being equal to "E-mail" but this is not
very portable between different locales.  Instead jpilot now uses the same scheme it uses
for exporting records(csv, ldif) which is to check for a labelcode of 4.  This is still a
hardcoded value but it seems to be the one Palm developers have chosen and is unlikely
to change anytime soon.