Stop printing contents of oversized note or description field in todo when field exceeds max length.
On entering an overly large note or description field jpilot emits a warning
and truncates the buffer to a maximum length.  In addition the current behavior
has been to print the entire problem buffer which obscures what is happening
and makes it look like there has been some weird corruption within the Jpilot program.