Improve code efficiency of Glists
Doubly-linked Glists are used throughout the code.
During their creation the inefficient g_list_append is used which requires
traversing the Glist.  Replaced g_list_append with g_list_prepend
The lists, which are delivered with a pointer to the tail, are
often traversed to find the head of the list.  By switching to g_list_prepend
the head of the list is already available and a second uneccessary list
traversal can be avoided.