Allow selection of text in weekview and monthview guis.
Code now uses the button_release event rather than button_press event to
switch from the weekview and monthly views to the jpilot calendar window.
This allows selections of text to be copied before the automatic
re-focus to the calendar widget.

This was done in response to a request on the jpilot-mailing list.