Major overhaul of sync code
When records conflict, i.e., modifications on both the Palm and the PC,
the previous behavior had been to overrule the Palm in all cases and
keep the PC record.  This can lead to unacceptable data loss.
The new strategy detects a conflict and keeps both records to allow the
user to merge changes manually.

Without writing equality comparison routines for each database it is
impossible to be completely sure that the records are different.  The
current code checks only to see if the records are the same length.
Changes which preserve length, such as SMITH to smith or 1234 to 4567,
will not be caught.  Nevertheless, the probabilistic chance of data loss
has decreased significantly since both a length-preserving change AND
a simultaneous modification of the same record is rare.  It is rare enough
that the previous unacceptable behaviour has not generated any comments
until recently.

Other problems fixed:
syncing without forking now works
a slow sync where the pdb database does not exist no longer locks the pc3 file