Validate repeating event start dates to prevent problems.
Users have encountered problems when the start date of a repeating event
is not, in fact, an actual date on which the event occurs.  For example, the
search dialog will find an appointment but clicking to go to that day will
not show anything.  This fix modifies the start date to be a valid one.