Searches which return datebook events now position the calendar on the next occurrence of an event.
This solves a problem where the pdb database has a repeating event which
starts on one day but actually has no occurrence on the start date.

This also changes the UI behavior slightly.  A search for a monthly
repeating event will now take you to the next occurrence rather than
the start date of the event which may have been years back.
This is useful because one is typically concerned with what other
irregular appts., such as a doctor's visit, might interfere with the
regularly scheduled appointments.

The search GUI has not been modified which may be potentially confusing.
It shows the start date of a repeating appt. but clicking on it will take
you to a different date than the one shown(i.e., the one closest to the
present).