Blank ContactDB not being copied correctly.
A new user (no sync) has the initial database files copied from the files
in /usr/share/jpilot.  The code to do the copying was writing an extra
EOF character at the end of the file so the two files were not identical.
This caused no problems with the old databases but was causing a problem
with the new Contacts database.

The old code for copying files may appear elsewhere in the code and
be causing unspecified errors.