Should fix bug 1918 - A Contacts database with a "version 10" appinfo structure is in the empty directory and I believe most Palms have version 11.
The T/Es have a version 10.

An app info struct that was too short (version 10 size) was being written
to the Palm and would cause it to need to be hard reset.

I tested this with my Treo 680, however I did not test it before I made this
change to verify the need to hard reset.

We might want to have a version 11 database in the empty directory, but I
want to test this fix with a version 10.