Temporarily remove button to edit categories for 4 main apps
PDAs are for saving data and time.  Unfortunately the category sync code often loses data and results in lost time trying to recreate the original database.  The simplest remedy is not to offer the edit category button and have users change categories on the Palm.

I have not changed libplugin.c so there is still a backdoor by which plugins like Expense can use the faulty code.