plugin_gui_cleanup(): only store the time if the plugin was active.
If the keyring plugin is active and a synch is performed then
plugin_gui_cleanup() is called. If the keyring plugin is called again
then plugin_gui_cleanup() is called again just before plugin_gui() is
called so the time difference is always < PLUGIN_MAX_INACTIVE_TIME and
no password is requested.