Simplify KeyRing host/Palm string handling.
KeyRing now translates strings from Palm to host character set at the
"boundaries" of the application.  In other words, it only translates once
when reading from a Palm file and once more when it writes back to a Palm
file.  This matches the behavior of the other applications and also yields
a performance benefit as unnecessary conversions are eliminated.

Export of KeyRing data, which motivated this patch, is now correctly in
the host character set.