Center pixmaps in clist columns below pixmap in title row of clist
Previously the pixmaps were not centered under their title icons but were
off to the side by 1 pixel.  This was small but visually annoying.  The
problem was two-fold.  A) the option to center the columns was not being
applied and B) GTK cannot center icons which have a width which is an odd number.