Add support for Location field in Calendar application
Implementation relies on pack/unpack routines that were borrowed from
pilot-link and amended to support the Calendar app.  These routines
in jp-calendar.c should eventually be migrated to pilot-link.

Currently categories are only supported on the right-hand side of the GUI.
The code could be changed to resemble the other apps and have a category
menu on the left as well.  The GUI might become a little too busy at that
point but I am open to a request from users for such a feature.